steps 1 to 2 of individual expand inventory
