+++ src/css/style.css
@@ -224,3 +224,8 @@ footer {
   font-size: var(--small-font);
   padding: 1em;
 }
+
+/* additional styles */
+.category_name {
+  text-align: center;
+}
