+++ src/index.html
@@ -30,9 +30,33 @@
           comfortably at an affordable price.
         </p>
       </section>
-      <section class="products">
-        <h2>Top Products</h2>
-        <ul class="product-list"></ul>
+      <section class="categories">
+        <ul class="product-list">
+          <li class="product-card">
+            <a href="product-listing/index.html?category=tents">
+              <img src="https://byui-cit.github.io/wdd330/images/category-icons/noun_Tent_3544767.svg" alt="Tent Icon from Noun Project: Mustofa Bayu">
+              <h2 class="card_name category_name">Tents</h2>  
+            </a>
+          </li>
+          <li class="product-card">
+            <a href="product-listing/index.html?category=backpacks">
+              <img src="https://byui-cit.github.io/wdd330/images/category-icons/noun_carrier%20bag_3544771.svg" alt="Backpack Icon from Noun Project: Mustofa Bayu">
+              <h2 class="card_name category_name">Backpacks</h2>  
+            </a>
+          </li>
+          <li class="product-card">
+            <a href="product-listing/index.html?category=sleeping-bags">
+              <img src="https://byui-cit.github.io/wdd330/images/category-icons/noun_Sleeping%20Bag_3544775.svg" alt="Sleeping Bag Icon from Noun Project: Mustofa Bayu">
+              <h2 class="card_name category_name">Sleeping Bags</h2>  
+            </a>
+          </li>
+          <li class="product-card">
+            <a href="product-listing/index.html?category=hammocks">
+              <img src="https://byui-cit.github.io/wdd330/images/category-icons/noun_Hammock_791143.svg" alt="Hammock Icon from Noun Project: Paul Richard">
+              <h2 class="card_name category_name">Hammocks</h2>  
+            </a>
+          </li>
+        </ul>
       </section>
     </main>
     <footer id="main-footer"></footer>
